feat: router
- add go router package
- add string
- use material app router method in home
- move home to splash screen widget
- create router file
