+++ pubspec.yaml
@@ -10,6 +10,7 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.8
+  go_router: ^14.2.8
 
 dev_dependencies:
   flutter_test:
